Separated added CSV and XML Handlers for Single Responsibility
Introduced XML Processing
